requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.50.0
PyYAML>=5.4.0
opencv-python>=4.5.0
tensorboard>=2.4.0 